Move maintenance routes to be under admin
