+++ backend/src/flow44/api/admin.py
@@ -1,3 +1,5 @@
+import asyncio
+import os
 from typing import Annotated
 
 from fastapi import APIRouter, Depends, HTTPException
@@ -5,11 +7,16 @@ from pydantic import BaseModel
 from sqlalchemy.exc import IntegrityError
 
 from flow44.api.deps import UserDep, is_admin
+from flow44.config import settings
 from flow44.db.platform_user import (
     add_platform_user,
     list_platform_users,
     remove_platform_user,
 )
+from flow44.db.project import Project
+from flow44.services.maintenance.runner import ProjectMaintenanceResult, run_over_projects
+from flow44.services.maintenance.sandbox_file_permissions import fix_file_permissions
+from flow44.services.maintenance.template_sync import sync_protected_template_files
 
 router = APIRouter(prefix="/api/admin", tags=["admin"])
 
@@ -53,3 +60,25 @@ async def revoke_user(user_id: AdminDep, target_user_id: str) -> None:
     removed = await remove_platform_user(target_user_id)
     if not removed:
         raise HTTPException(status_code=404, detail="User not found")
+
+
+@router.post("/maintenance/fix-file-permissions", tags=["maintenance"])
+async def fix_workspace_file_permissions(user_id: AdminDep) -> list[ProjectMaintenanceResult]:
+    """Make all workspace files world-writable. (new files will be world-writable already after os.umask(0))"""
+
+    async def fix_permissions(project: Project) -> str:
+        workspace_dir = os.path.join(settings.WORKSPACE_BASE_DIR, project.id)
+        return await asyncio.to_thread(fix_file_permissions, workspace_dir)
+
+    return await run_over_projects(fix_permissions)
+
+
+@router.post("/maintenance/sync-protected-files", tags=["maintenance"])
+async def sync_protected_files(user_id: AdminDep) -> list[ProjectMaintenanceResult]:
+    """Overwrite each project's template files with the up-to-date template if they differ."""
+
+    async def sync_files(project: Project) -> str:
+        workspace_dir = os.path.join(settings.WORKSPACE_BASE_DIR, project.id)
+        return await asyncio.to_thread(sync_protected_template_files, workspace_dir, settings.TEMPLATE_DIR)
+
+    return await run_over_projects(sync_files)
